move some common method to abstract class
